Final Commit
Ready for review. Inserted a video from YouTube using the iframe element and also added two different audio clips using the audio element with controls. There is also an embed element containing the audio for browsers that support plug-ins, and there is a message for those who's browsers don't support plug-ins or HTML5.